Core/World/VMAPFactory: Add new config setting LoadAllGridsOnMaps
With this setting one can store a list of maps that are automatically
loaded with all grids and all creatures.
Continents will be loaded on server-startup, instances when a player
enters.

NOTE: Using maps with this setting will cost huge amount of cpu and ram
power, and hence it must only be used for development purposes